fix: Allow floating point BPM values
The heart-rate API was previously failing when receiving floating point numbers for BPM values. This was due to a strict integer validation in the schema.

This commit updates the validation schema to allow floating point numbers for BPM values, which makes the API more robust to different sensor data formats.

+++ app/api/heart-rate/route.ts
@@ -17,7 +17,7 @@ const IncomingHeartRateSchema = z.object({
 const ValidatedHeartRateData = z.object({
   username: z.string().min(1),
   timestamp: z.array(z.string().datetime({ offset: true })),
-  bpm: z.array(z.number().int()),
+  bpm: z.array(z.number()),
 });
 
 interface HeartRateEntry {
